+++ src/scrapers/ScraperInfos.h
@@ -39,8 +39,9 @@ enum class MovieScraperInfo : int
     ClearArt = 23,
     Banner = 24,
     Thumb = 25,
+    TvShowLink = 26, // not used for scraping
     First = 1,
-    Last = 25
+    Last = 26
 };
 
 namespace mediaelch {